API/requirements.txt
tensorflow>=2.16.0
tensorflow-serving-api>=2.16.0
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy<2.0.0
